Move types.RepoCloneURL to repos.CloneURL
The code is related to repos, it should belong in the repos pacakge.
Also, havning it in the types package causaed that package to pull in a
bunch of code host specific packages.
