+++ internal/repos/syncer_test.go
@@ -98,7 +98,7 @@ func testSyncerSync(t *testing.T, s *repos.Store) func(*testing.T) {
 			ServiceID:   "https://git-host.com/",
 			ServiceType: extsvc.TypeOther,
 		},
-		Metadata: &types.OtherRepoMetadata{},
+		Metadata: &extsvc.OtherRepoMetadata{},
 	}).With(
 		types.Opt.RepoSources(otherService.URN()),
 	)
